requirements.txt
tensorflow==2.13.1
pandas
matplotlib
wandb
Flask==3.0.2
gunicorn==21.2.0
numpy
opencv-python-headless==4.9.0.80
Pillow